GO-4459: Rename service dir to internal
