+++ core/api/internal/search/handler.go
