+++ core/api/internal/space/model.go
