+++ core/api/server/server.go
@@ -6,11 +6,11 @@ import (
 	"github.com/gin-gonic/gin"
 
 	"github.com/anyproto/anytype-heart/core/anytype/account"
-	"github.com/anyproto/anytype-heart/core/api/services/auth"
-	"github.com/anyproto/anytype-heart/core/api/services/export"
-	"github.com/anyproto/anytype-heart/core/api/services/object"
-	"github.com/anyproto/anytype-heart/core/api/services/search"
-	"github.com/anyproto/anytype-heart/core/api/services/space"
+	"github.com/anyproto/anytype-heart/core/api/internal/auth"
+	"github.com/anyproto/anytype-heart/core/api/internal/export"
+	"github.com/anyproto/anytype-heart/core/api/internal/object"
+	"github.com/anyproto/anytype-heart/core/api/internal/search"
+	"github.com/anyproto/anytype-heart/core/api/internal/space"
 	"github.com/anyproto/anytype-heart/pb/service"
 )
